DBC: rename `dbc.py` to `dbc_loader.py` and outsource the writing code to `dbc_writer.py`
Approved-by: Alexander Walz <alexander.walz@mercedes-benz.com>
